requirements.txt
cbbd>=1.21.0
pandas==2.1.4
numpy==1.26.2
scikit-learn==1.3.2
python-dotenv==1.0.0
tabulate==0.9.0

